Initial Commit
My submission at the end of the working day. Managed to get many features fleshed out , need a bit more time to finish. Regardless as a learning tool i have learnt Binding in MVVM aswell as practiced how to properly control interface and model integration and the flow of data.